+++ fuc/cli/bam_aldepth.py
@@ -7,6 +7,8 @@ description = f"""
 # Count allelic depth from a SAM/BAM/CRAM file. #
 #################################################
 
+The 'sites' file can be a TSV file containing two columns, chromosome and position. It can also be a BED or VCF file. Input file type will be detected automatically.
+
 Usage examples:
   $ fuc {api.common._script_name()} in.bam sites.tsv > out.tsv
   $ fuc {api.common._script_name()} in.bam sites.bed > out.tsv
@@ -26,7 +28,7 @@ def create_parser(subparsers):
     )
     parser.add_argument(
         'sites',
-        help='TSV file containing two columns, chromosome and position. Also accepts a BED or VCF file.'
+        help='TSV/BED/VCF file (zipped or unzipped).'
     )
 
 def main(args):